pyproject.toml
[project]
name = "nus-hackathon-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.135.2",
    "greenlet>=3.3.2",
    "minio>=7.2.20",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
